update soter/openssl to fix compatibility with new openssl version (#258)
* use pointers instead stack structs

* fix usage of openssl hash and rsa structs

* update openssl ecdsa sign usage

* fix test running with cmake

* use explicit initializing digest context with new openssl api

* CMakeLists.txt simple non-productions script to build themis/soter/tests (for ide)

* fix secure_session tests
mark unused files to fix "ISO C forbids an empty source file" warning

* fix valgrind check

* fix memory leaks and errors in tests

* fix leak in old openssl after new changes

* install last valgrind on ci
turn off one nist test on ci by cflags

* fix installing valgrind

* handle stack structs

* cache valgrind/golang fetching in ci

* add valgrind dependency

* restore installing golang

* test cache on ci

* fix check cache

* revert using calloc instead malloc
fix mistake of merging test.mk

* revert themis_secure_session.c test changes

* run nist test via cmakelists

* use stack allocation of soter_hash instead heap in secure_comparator (revert prev changes)

* Revert "revert themis_secure_session.c test changes"

This reverts commit 6913bdff62e9d05c862e81b4b446d54b13e57c90.

* run themis tests with boringssl and valgrind

* fix memory leak in boringssl/soter_hash

* drop openssl1.0.2 dependency for stretch

* add failure check
fix incorrect indentation

+++ src/soter/boringssl/soter_sign_rsa.c
@@ -101,8 +101,8 @@ soter_status_t soter_sign_final_rsa_pss_pkcs8(soter_sign_ctx_t* ctx, void* signa
   if (!pkey){
     return SOTER_INVALID_PARAMETER;
   }
-  if((*signature_length)<EVP_PKEY_size(pkey)){
-    (*signature_length)=EVP_PKEY_size(pkey);
+  if((*signature_length)<(size_t)EVP_PKEY_size(pkey)){
+    (*signature_length)=(size_t)EVP_PKEY_size(pkey);
     return SOTER_BUFFER_TOO_SMALL;
   }
   if(!EVP_DigestSignFinal(ctx->md_ctx, signature, signature_length)){
